Move common proof interfaces to the parent package
to improve discoverability
