+++ exonum-java-binding-common/src/main/java/com/exonum/binding/common/proofs/map/CheckedMapProof.java
@@ -16,7 +16,7 @@
 
 package com.exonum.binding.common.proofs.map;
 
-import com.exonum.binding.common.proofs.common.CheckedProof;
+import com.exonum.binding.common.proofs.CheckedProof;
 import com.google.protobuf.ByteString;
 import java.util.Set;
